feat(sso-login): add automatic initialisation of login windows
add possibility to pass an optional "autoLogin" parameter. if this parameter
is set, the login process is automatically started as soon as the providers
are loaded.
