+++ packages/sso-login/src/components/LoginBox/LoginBox.js
@@ -1,12 +1,22 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
+import {openLoginWindow} from '../../utils/loginWindow'
 import ProviderButton from '../ProviderButton/ProviderButton'
 
 class LoginBox extends React.Component {
   constructor(props) {
     super(props)
     props.loadProviders()
+  }
+
+  componentDidUpdate(prevProps, prevState, snapshot) {
+    if (this.props.autoLogin && prevProps.providers.length === 0 && this.props.providers.length > 0) {
+      const autoLoginProvider = this.props.providers.find(entity => entity.unique_id === this.props.autoLogin)
+      if (autoLoginProvider) {
+        openLoginWindow(this.props.loginEndpoint, this.props.loginCompleted, autoLoginProvider)
+      }
+    }
   }
 
   render() {
@@ -26,7 +36,8 @@ LoginBox.propTypes = {
   loginEndpoint: PropTypes.string,
   loginCompleted: PropTypes.func.isRequired,
   loadProviders: PropTypes.func.isRequired,
-  providers: PropTypes.array
+  providers: PropTypes.array,
+  autoLogin: PropTypes.string
 }
 
 export default LoginBox
